uprava moduly, fce na test
