+++ 06 piškvorky/tick_tack_toe_with_tests/ai.py
@@ -1,5 +1,4 @@
 from random import randrange
-from piskvorky import move
 
 
 def pc_move(area):  # move for pc, position choose by randrange
@@ -8,6 +7,15 @@ def pc_move(area):  # move for pc, position choose by randrange
         position = randrange(0, 19)
         if area[int(position)] == "-":
             return move(area, position, pc_symbol)
+
+
+def move(area, position, symbol):
+    # put symbol into area, get area, position 1 - 19, symbol
+
+    area = list(area)
+    area[position] = symbol
+    area = ''.join(area)
+    return area
 
 
 pc_symbol = "O"
